+++ src/main/java/wooteco/subway/exception/validation/SectionNotSuitableException.java
@@ -1,4 +1,6 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.validation;
+
+import wooteco.subway.exception.SubwayValidationException;
 
 public class SectionNotSuitableException extends SubwayValidationException {
